+++ packages/integration-tests/fixtures/release-value-with-quotes/release-value-with-quotes.test.ts
@@ -6,20 +6,10 @@ import { testIfNodeMajorVersionIsLessThan18 } from "../../utils/testIf";
 
 function checkBundle(bundlePath: string): void {
   const output = execSync(`node ${bundlePath}`, { encoding: "utf-8" });
-
-  const map = JSON.parse(output) as Record<string, string>;
-
-  // There should be only one key in the map
-  expect(Object.values(map)).toHaveLength(1);
-  // The value should be the expected metadata
-  expect(Object.values(map)).toEqual([{ team: "frontend" }]);
-
-  // The key is the stack trace of the error thrown in the file
-  expect(Object.keys(map)[0]).toContain("Error");
-  expect(Object.keys(map)[0]).toContain("bundle.js");
+  expect(output.trimEnd()).toBe('"i am a dangerous release value because I contain a \\""');
 }
 
-describe("metadata injection", () => {
+describe("Properly escapes release values before injecting", () => {
   testIfNodeMajorVersionIsLessThan18("webpack 4 bundle", () => {
     checkBundle(path.join(__dirname, "out", "webpack4", "bundle.js"));
   });